rename experiment_management as ExperimentClient, rewrote ExperimentClient as a class, moved tests into separate untracked file
